🧬🧶 Slight refactor after revision
Why
- Ordinary routine revision

How
- Added some WIP basic descriptions to the 'utils' functions
- Refactored the 'utils' functions after revision
- Renamed 'calendar' store to 'calendarStore'

Tags
 Changes to be committed:
	modified:   src/components/Days.vue
	modified:   src/components/utils.ts
	renamed:    src/stores/calendar.ts -> src/stores/calendarStore.ts
	modified:   src/views/Calendar.vue
